requirements.txt
python_dotenv==1.0.1
ollama==0.4.4
